Turn off gzip encoding for Enki
Full disclaimer: this might be addressing a symptom rather than the
disease, but it works. If the CM needs to get the content for an Enki
book that is not DRM-protected, it will explicitly ask for it with
any content encoding. The clients should be able to handle gzip
encoding, but for some reason iOS was unable to download these books.