Add version to heartbeat endpoint
Using the Internet-Draft "Health Check Response Format for HTTP APIs"
(https://inadarei.github.io/rfc-healthcheck/), this commit adds the application
version to the heartbeat controller response.

This endpoint is intended to support administrative insight into the system
and may be extended later with the current content of the /service_status
endpoint.