Add Dockerfile.exec for script execution
Images created with this Dockefile will execute a single script on the
Library Simplified Circulation Manager, and then stop running.

This image is an alternative and/or supporting option when compared with
Dockefile.scripts images, which attempts to run all scripts (required
and optional) in an ongoing cron service.

What this image lacks in automation, it makes up for in configurability.
It is intended for use in advanced deployment configurations.

Resolves NYPL-Simplified/circulation-docker#77 (GitHub).