Make SIP Connections short lived
LYRASIS was seeing issues using SIP2 authentication with the
Yavapai library network where a large number of SIP processes
were being created in the ILS (SirsiDynix Symphony) which was
causing performce problems in the ILS.

This patch updates the SIP client to:
  - Connect Socket
  - Send Login message - 93 (if necessary)
  - Send Patron Information message - 63
  - Send End Patron Session - 35
  - Close Socket

There is no attempt at long lived connections or socket reuse,
this process is done every time. Based on conversations with
Yavapai this appears to be how Overdrives SIP2 authentication
works.