Some updates to the patron web hostname settings
In #1447 there were some issues with how the patron web hostname
settings were saved. They are not saved as json, they are saved in
text, so there are issues when we try to load them as json.

This PR backs out those changes, and instead saves the hostnames
to the DB as a comma seperated list and breaks them out when the
settings are loaded.