Fix issue with uszipcode library not loading DB
The uszipcode library downloads a sql file on first request that it
uses to look up zipcodes. The host of this file is down currently and
appears to have been for several days.

This PR takes the strategy outlined here:
https://github.com/MacHu-GWU/uszipcode-project/issues/40#issuecomment-688323810

It puts the db file in the repo, and passes the location of the DB into
the uszipcode library so it doesn't have to go out and download it.