Disable Tox hash randomization for now.
Tox helpfully sets the PYTHONHASHSEED value to a random number to make
sure that tests do not depend on the order of data returned from
iterating over hashes. This is the default behavior of Python from
verison 3.3. See [1].

There are places in the tests where we do expect an explicit order when
iterating over values from a hash. These will need to be addressed, but
for now just disable this PYTHONHASHSEED randomization.

[1] - https://tox.readthedocs.io/en/latest/example/basic.html#special-handling-of-pythonhashseed