Replace travisci with github actions
This removes travis CI, and replaces it with a combination of tox and
GH actions. In the future this will easily allow us to test with multiple
versions of Python as we move towards supporting Python 2 and Python 3
at the same time.

It should also allow us to run these test locally.