Break requirements out into 3 files.
One for development, one for production and one containing the common
components for each of them.

Having seperate requirements for dev and production lets us only install
the packages used for testing into the test environment.

The reason we have a common "base" requirements file is that the
psycopg2 package has two versions (psycopg2-binary and psycopg2)[1][2].
The -binary version isn't recommended for production use, but is very
helpful for testing, so we don't have to build the psycopg2 package
every time CI runs.

[1] - https://github.com/psycopg/psycopg2/issues/674
[2] - https://www.psycopg.org/articles/2018/02/08/psycopg-274-released/