Update test_opds_import.py
Use some pytest fixtures to patch up the tests in test_opds_import.py
to work around the non-deterministic behavior exhibited by the OPDS
importer when the PYTHONHASHVALUE is changed.

The importer iterates over a hash, so the order that things are imported
in changes. The tests need to take this into account.