WIP OpenAPI documentation for /admin/* endpoints
This is an early working draft of adding OpenAPI support to the circulation-manager endpoints. It includes documentation for some endpoints as an example and can be extended to cover all API endpoints with the addition of further docstrings (and relevant schemas).

There is room for a lot of improvement in the code:

- the OpenAPI spec is generated for each request. This should be generated when the container comes up and stored somewhere. (there are circular dependencies that make this difficult at the moment)
- The schemas generated currently happen in one large method. This could/should be refactored to be more manageable
- Authentication is not currently correctly configured

This is a test PR to solitit feedback and evaluate if this is a viable way of documenting the API structure