TEST Add CSRF Token to sample endpoint
This utilizes a plugin to fetch the CSRF token and add it to a sample API endpoint. This allows for `POST` requests to be made via this endpoint. This is not an ideal solution -- the token is visible to the user -- but it helps make clear how requests must be made.

This also does not solve the issue of how initial authentication, as of right now a user still must access the `/admin` endpoint first to generate a CSRF token before it can be populated here.