BUGFIX Switch Container to bitnami/elasticsearch
This swaps out the Elastic.co container for one maintained by Bitnami. It includes some conveneince tooling that can install plugins, removing the need to define a custom command for the container. The bitnami maintained image closely tracks the officially supported images (up through 8.x at the time of this commit), so we do not foresee any issues with future support.