SIMPLY-4089 Add Unsubscribe link option
This adds an option for a library to have an (optional) link to unsubscribe to email notifications/communications. This is part of the `Patron Support` page and returns the URI as part of the `links` array as part of the `authentication_document`.

In addition this adds a new convenience method `_help_uri_with_fallback` that will return a help URI, falling back to the general help link configured for the library. This compliments the `_email_uri_with_fallback` method.

In addition one test is updated to reflect the new state of the `authentication_document`.