SIMPLY-4113 Remove app-specific config from INI
To ensure that the CM remains deployable by all institutions this removes any naming and monitoring settings for NewRelic from the INI file, requiring that they be set as environment variables. These are set in the test Docker Compose files, which can be changed to enable local debugging of NewRelic configuration (or renaming the application in an individual NewRelic instance).