NOREF Close session from Config.cdns method
In the Config module there is a specific method that generates its own database session in local scope. This is used to check for and load CDN configuration and is potentially called for non-cached XML feeds for cover image fetching.

This session object is never closed/removed, which could be contributing to persistent database connection issues. These have been found across CM instances. This may or may not be implicated, regardless database sessions should be explicitly closed after opening.