NOREF Update Error Code from Overdrive
Following an error report from a partner, it has come to our attention that an error returned from the Overdrive API is not handled correctly. This update simply adds the correct error code, which will result in responses containing that error setting available copies to zero. This should have no other effects