NOREF Update update_lane_size script schedule
This is an expensive operation which can be taxing on the database, it should be run earlier in the day when there is less patron activity. This should have a two-fold impact: 1) It should improve the script's speed and 2) it should reduce the impact of any associated perfromance issues as we see lower traffic at that time.