SMA-523 Add since/until params for open sources materials
In order to allow libraries to set an expected expiration date for open source materials this adds `opds:since` and `opds:until` to the availability object for these materials. This leverages functionality that is already used for loans to create a "loan-like" behavior for these materials (when desired).

This should have no side effects since it's leveraging existing attributes but this should be tested to be true.